+++ src/main/java/main/algorithms/impls/BittorrentAlgorithmInitializer.java
@@ -11,8 +11,10 @@ import main.file.system.FileSystemLink;
 import main.file.system.allocator.AllocatorStore;
 import main.peer.Link;
 import main.peer.peerMessages.PeerMessage;
+import main.peer.peerMessages.RequestMessage;
 import main.torrent.status.TorrentStatusAction;
 import main.torrent.status.state.tree.TorrentStatusState;
+import reactor.core.publisher.EmitterProcessor;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.FluxSink;
 import reactor.core.publisher.UnicastProcessor;
@@ -25,7 +27,7 @@ public class BittorrentAlgorithmInitializer {
                                          TorrentInfo torrentInfo,
                                          Store<TorrentStatusState, TorrentStatusAction> store,
                                          FileSystemLink fileSystemLink,
-                                         UnicastProcessor<AbstractMap.SimpleEntry<Link, PeerMessage>> incomingPeerMessages$,
+                                         EmitterProcessor<AbstractMap.SimpleEntry<Link, PeerMessage>> incomingPeerMessages$,
                                          FluxSink<AbstractMap.SimpleEntry<Link, PeerMessage>> emitIncomingPeerMessages,
                                          Flux<Link> peersCommunicatorFlux,
                                          String identifier) {
@@ -45,14 +47,14 @@ public class BittorrentAlgorithmInitializer {
         UploadAlgorithm uploadAlgorithm = new UploadAlgorithmImpl(torrentInfo,
                 store,
                 fileSystemLink,
-                incomingPeerMessages$,
-                peersCommunicatorFlux);
+                incomingPeerMessages$.filter(tuple2 -> tuple2.getValue() instanceof RequestMessage)
+                        .map(tuple2 -> new AbstractMap.SimpleEntry<>(tuple2.getKey(), (RequestMessage) (tuple2.getValue()))));
 
         PeersToPiecesMapper peersToPiecesMapper =
                 new PeersToPiecesMapperImpl(recordedPeerFlux,
                         fileSystemLink.getUpdatedPiecesStatus());
 
-        BlockDownloader blockDownloader = new BlockDownloaderImpl(torrentInfo, fileSystemLink,identifier);
+        BlockDownloader blockDownloader = new BlockDownloaderImpl(torrentInfo, fileSystemLink, identifier);
 
         PiecesDownloader piecesDownloader = new PiecesDownloaderImpl(allocatorStore,
                 torrentInfo, store,
