+++ src/main/java/main/algorithms/impls/v1/upload/UploadAlgorithmImpl.java
@@ -6,9 +6,6 @@ import main.downloader.PieceEvent;
 import main.downloader.TorrentPieceStatus;
 import main.file.system.FileSystemLink;
 import main.peer.Link;
-import main.peer.ReceiveMessagesNotifications;
-import main.peer.peerMessages.PeerMessage;
-import main.peer.peerMessages.PieceMessage;
 import main.peer.peerMessages.RequestMessage;
 import main.torrent.status.TorrentStatusAction;
 import main.torrent.status.state.tree.TorrentStatusState;
@@ -16,13 +13,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.publisher.UnicastProcessor;
 import reactor.core.scheduler.Scheduler;
 import reactor.core.scheduler.Schedulers;
 import redux.store.Store;
 
 import java.util.AbstractMap;
-import java.util.Map;
 
 public class UploadAlgorithmImpl implements UploadAlgorithm {
     private static Logger logger = LoggerFactory.getLogger(UploadAlgorithmImpl.class);
@@ -34,8 +29,7 @@ public class UploadAlgorithmImpl implements UploadAlgorithm {
                                Store<TorrentStatusState, TorrentStatusAction> store,
                                FileSystemLink fileSystemLink,
                                // TODO: I need only request messages here:
-                               UnicastProcessor<AbstractMap.SimpleEntry<Link, PeerMessage>> incomingPeerMessages$,
-                               Flux<Link> peersCommunicatorFlux) {
+                               Flux<AbstractMap.SimpleEntry<Link, RequestMessage>> incomingRequestMessages$) {
 
         Flux<TorrentStatusState> startUpload$ = store.statesByAction(TorrentStatusAction.START_UPLOAD_IN_PROGRESS)
                 .concatMap(__ -> store.dispatch(TorrentStatusAction.START_UPLOAD_SELF_RESOLVED))
@@ -50,23 +44,21 @@ public class UploadAlgorithmImpl implements UploadAlgorithm {
                 the solution is to report that I'm ready to upload only after I subscribed to the requests stream.
                 */
                 .concatMap(__ -> {
-                    Flux<PieceEvent> uploader$ = incomingPeerMessages$.filter(tuple2 -> tuple2.getValue() instanceof RequestMessage)
-                            .map(tuple2 -> new AbstractMap.SimpleEntry<Link, RequestMessage>(tuple2.getKey(), (RequestMessage) (tuple2.getValue())))
-                            .flatMap(tuple2 -> {
-                                Link link = tuple2.getKey();
-                                return Mono.just(tuple2.getValue())
-                                        .doOnNext(requestMessage -> logger.debug("App uploader - start analyze request: " + requestMessage))
-                                        .doOnNext(requestMessage -> logger.debug("App uploader - do I have the piece: " + requestMessage + " => " +
-                                                fileSystemLink.havePiece(requestMessage.getIndex())))
-                                        .filter(requestMessage -> fileSystemLink.havePiece(requestMessage.getIndex()))
-                                        .flatMap(requestMessage -> store.notifyWhen(TorrentStatusAction.RESUME_UPLOAD_WIND_UP, requestMessage))
-                                        .doOnNext(requestMessage -> logger.debug("start creating piece-message for response to peer because he sent me request-message: " + requestMessage))
-                                        .flatMap(fileSystemLink::buildPieceMessage)
-                                        .doOnNext(requestMessage -> logger.debug("end creating piece-message for response to peer because he sent me request-message: " + requestMessage))
-                                        .publishOn(UploadAlgorithmImpl.uploaderScheduler)
-                                        .flatMap(pieceMessage -> link.sendMessages().sendPieceMessage(pieceMessage)
-                                                .map(___ -> new PieceEvent(TorrentPieceStatus.UPLOADING, pieceMessage)));
-                            })
+                    Flux<PieceEvent> uploader$ = incomingRequestMessages$.flatMap(tuple2 -> {
+                        Link link = tuple2.getKey();
+                        return Mono.just(tuple2.getValue())
+                                .doOnNext(requestMessage -> logger.debug("App uploader - start analyze request: " + requestMessage))
+                                .doOnNext(requestMessage -> logger.debug("App uploader - do I have the piece: " + requestMessage + " => " +
+                                        fileSystemLink.havePiece(requestMessage.getIndex())))
+                                .filter(requestMessage -> fileSystemLink.havePiece(requestMessage.getIndex()))
+                                .flatMap(requestMessage -> store.notifyWhen(TorrentStatusAction.RESUME_UPLOAD_WIND_UP, requestMessage))
+                                .doOnNext(requestMessage -> logger.debug("start creating piece-message for response to peer because he sent me request-message: " + requestMessage))
+                                .flatMap(fileSystemLink::buildPieceMessage)
+                                .doOnNext(requestMessage -> logger.debug("end creating piece-message for response to peer because he sent me request-message: " + requestMessage))
+                                .publishOn(UploadAlgorithmImpl.uploaderScheduler)
+                                .flatMap(pieceMessage -> link.sendMessages().sendPieceMessage(pieceMessage)
+                                        .map(___ -> new PieceEvent(TorrentPieceStatus.UPLOADING, pieceMessage)));
+                    })
                             .replay()
                             .autoConnect(0);
